Add a new ESLint rule for main.ts (#97)
* feat(eslint): added new `no-extra-functions-in-main` rule

* refactor: remove export from sanitizeFolderName function

* refact(scripts): refactored some script to reflect new main rules

* feat: implement SQS message dumping and import notification workflow with error handling and progress reporting
