+++ scripts/go/go-analyze-alarm/src/main.ts
@@ -5,72 +5,21 @@
  * selects the appropriate runbook, and executes it via the RunbookEngine.
  */
 
-import { CloudWatchLogsClient } from '@aws-sdk/client-cloudwatch-logs';
-import { CloudWatchClient } from '@aws-sdk/client-cloudwatch';
-import { AthenaClient } from '@aws-sdk/client-athena';
-import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
-import { fromIni } from '@aws-sdk/credential-provider-ini';
-
 import { Core, Runbook } from '@go-automation/go-common';
 
 import type { GoAnalyzeAlarmConfig } from './types/GoAnalyzeAlarmConfig.js';
 import { buildAddressBookIoApiGwAlarmRunbook } from './libs/runbooks/address-book-io-api-gw-alarm/index.js';
 import { buildDeliveryB2BApiGwAlarmRunbook } from './libs/runbooks/delivery-b2b-api-gw-alarm/index.js';
 import { DEFAULT_TIME_WINDOW_MINUTES } from './libs/runbooks/address-book-io-api-gw-alarm/constants.js';
+import { createServiceRegistry } from './libs/createServiceRegistry.js';
+import { computeTimeRange } from './libs/computeTimeRange.js';
+import { saveExecutionTrace } from './libs/saveExecutionTrace.js';
 
 /** Runbook registry: maps alarm names to their runbook builders */
 const RUNBOOK_REGISTRY = new Map<string, () => Runbook.Runbook>([
   ['pn-address-book-io-IO-ApiGwAlarm', buildAddressBookIoApiGwAlarmRunbook],
   ['pn-delivery-B2B-ApiGwAlarm', buildDeliveryB2BApiGwAlarmRunbook],
 ]);
-
-/**
- * Creates a ServiceRegistry from an AWS SSO profile.
- *
- * @param profile - AWS SSO profile name
- * @returns ServiceRegistry with all services initialized
- */
-function createServiceRegistry(profile: string): Runbook.ServiceRegistry {
-  const credentials = fromIni({ profile });
-  const region = 'eu-south-1';
-
-  const cloudWatchLogsClient = new CloudWatchLogsClient({ region, credentials });
-  const cloudWatchClient = new CloudWatchClient({ region, credentials });
-  const athenaClient = new AthenaClient({ region, credentials });
-  const dynamoDBClient = new DynamoDBClient({ region, credentials });
-
-  return {
-    cloudWatchLogs: new Runbook.CloudWatchLogsService(cloudWatchLogsClient),
-    cloudWatchMetrics: new Runbook.CloudWatchMetricsService(cloudWatchClient),
-    athena: new Runbook.AthenaService(athenaClient, 's3://placeholder-athena-results/'),
-    dynamodb: new Runbook.RunbookDynamoDBService(dynamoDBClient),
-    http: new Runbook.RunbookHttpService(),
-  };
-}
-
-/**
- * Computes the time range from the alarm datetime string.
- * Returns start/end ISO strings with ±timeWindowMinutes.
- *
- * @param alarmDatetime - ISO 8601 timestamp of the alarm
- * @param timeWindowMinutes - Time window in minutes
- * @returns Start and end ISO strings
- */
-function computeTimeRange(alarmDatetime: string, timeWindowMinutes: number): { startTime: string; endTime: string } {
-  const alarmTime = new Date(alarmDatetime);
-  if (Number.isNaN(alarmTime.getTime())) {
-    throw new Error(`Invalid alarm datetime: "${alarmDatetime}". Expected ISO 8601 format.`);
-  }
-
-  const offsetMs = timeWindowMinutes * 60 * 1000;
-  const start = new Date(alarmTime.getTime() - offsetMs);
-  const end = new Date(alarmTime.getTime() + offsetMs);
-
-  return {
-    startTime: start.toISOString(),
-    endTime: end.toISOString(),
-  };
-}
 
 /**
  * Main script execution function.
@@ -156,27 +105,4 @@ export async function main(script: Core.GOScript): Promise<void> {
 
   // Save execution trace to data directory
   await saveExecutionTrace(script, result, config.alarmName);
-}
-
-/**
- * Saves the RunbookExecutionTrace as a JSON file in the script's data directory.
- * File name: trace-{alarmName}-{timestamp}.json
- *
- * @param script - The GOScript instance for path resolution
- * @param result - The runbook execution result containing the trace
- * @param alarmName - Alarm name used in the file name
- */
-async function saveExecutionTrace(
-  script: Core.GOScript,
-  result: Runbook.RunbookExecutionResult,
-  alarmName: string,
-): Promise<void> {
-  const fileName = `trace-${alarmName}.json`;
-  const traceInfoPath = script.paths.resolvePathWithInfo(fileName, Core.GOPathType.OUTPUT);
-  const tracePath = traceInfoPath.path;
-
-  const exporter = new Core.GOJSONFileExporter({ outputPath: tracePath });
-  await exporter.export(result.trace);
-
-  script.logger.info(`Execution trace saved: ${tracePath}`);
 }
