+++ scripts/send/send-monitor-tpp-messages/src/main.ts
@@ -5,183 +5,24 @@
  * Receives typed dependencies (script) for clean separation of concerns.
  */
 
-import * as path from 'path';
-
-import { DateTime } from 'luxon';
-
 import { Core } from '@go-automation/go-common';
 
 import {
   AwsAthenaService,
   AthenaQueryExecutor,
   convertAthenaResults,
-  analyzeThreshold,
-  generateThresholdReport,
   SlackNotifier,
-  formatDateForAthena,
-  getDateComponents,
   parseDateTime,
   hoursAgo,
 } from './libs/index.js';
+import { buildQueryParams } from './libs/buildQueryParams.js';
+import { saveAndAnalyzeResults } from './libs/saveAndAnalyzeResults.js';
+import { sendSlackReport, notifySlackError } from './libs/sendSlackReport.js';
 import type { AthenaQueryConfig } from './types/AthenaQueryConfig.js';
-import type { AthenaQueryResults } from './types/AthenaQueryResults.js';
-import type { CSVRow } from './types/CSVRow.js';
-import type { QueryParams } from './types/QueryParams.js';
 import type { TPPMonitorConfig } from './types/TPPMonitorConfig.js';
 
-// ============================================================================
-// Helper Functions
-// ============================================================================
-
 /**
- * Builds query parameters from parsed date range
- *
- * @param startDate - Start date for query
- * @param endDate - End date for query
- * @returns Query parameters object
- */
-function buildQueryParams(startDate: Date, endDate: Date): QueryParams {
-  const startComponents = getDateComponents(startDate);
-  const endComponents = getDateComponents(endDate);
-
-  return {
-    startDate: formatDateForAthena(startDate),
-    endDate: formatDateForAthena(endDate),
-    startYear: startComponents.year,
-    startMonth: startComponents.month,
-    startDay: startComponents.day,
-    startHour: startComponents.hour,
-    endYear: endComponents.year,
-    endMonth: endComponents.month,
-    endDay: endComponents.day,
-    endHour: endComponents.hour,
-  };
-}
-
-/**
- * Executes the Athena query and returns results
- *
- * @param executor - Athena query executor instance
- * @param queryTemplate - SQL query template
- * @param athenaConfig - Athena configuration
- * @param queryParams - Query parameters
- * @returns Query results
- */
-async function executeAthenaQuery(
-  executor: AthenaQueryExecutor,
-  queryTemplate: string,
-  athenaConfig: AthenaQueryConfig,
-  queryParams: QueryParams,
-): Promise<AthenaQueryResults> {
-  return executor.executeQuery(queryTemplate, athenaConfig, queryParams);
-}
-
-/**
- * Generates a timestamped filename for the CSV report
- *
- * @param prefix - Filename prefix
- * @returns Generated filename with timestamp
- */
-function generateFileName(prefix: string = 'report'): string {
-  const timestamp = DateTime.now().toFormat('yyyy-MM-dd_HH-mm-ss');
-  return `${prefix}_${timestamp}.csv`;
-}
-
-/**
- * Saves results to CSV using GOCSVListExporter and analyzes them
- *
- * @param data - Converted CSV rows
- * @param outputFolder - Directory for CSV output (resolved by GOPaths)
- * @param thresholdField - Optional field to analyze
- * @param threshold - Threshold value
- * @returns Analysis results with file path and report
- */
-async function saveAndAnalyzeResults(
-  data: CSVRow[],
-  outputFolder: string,
-  thresholdField?: string,
-  threshold?: number,
-): Promise<{
-  csvFilePath: string | null;
-  fileName: string | null;
-  rowCount: number;
-  analysis: string;
-}> {
-  const rowCount = data.length;
-
-  let csvFilePath: string | null = null;
-  let fileName: string | null = null;
-
-  if (rowCount > 0) {
-    fileName = generateFileName();
-    csvFilePath = path.join(outputFolder, fileName);
-
-    const exporter = new Core.GOCSVListExporter<CSVRow>({
-      outputPath: csvFilePath,
-      includeHeader: true,
-    });
-    await exporter.export(data);
-  }
-
-  let analysis: string;
-  if (rowCount === 0) {
-    analysis = 'No data found in the specified time range';
-  } else if (thresholdField && threshold !== undefined && threshold > 0) {
-    const flaggedRows = analyzeThreshold(data, thresholdField, threshold);
-    analysis = generateThresholdReport(flaggedRows, thresholdField, threshold);
-  } else {
-    analysis = `Found ${rowCount} rows`;
-  }
-
-  return { csvFilePath, fileName, rowCount, analysis };
-}
-
-/**
- * Sends report to Slack if configured
- *
- * @param slackNotifier - Slack notifier instance (or null)
- * @param messageTemplate - Message template
- * @param startDate - Query start date
- * @param endDate - Query end date
- * @param csvFilePath - Path to CSV file (or null)
- * @param rowCount - Number of rows
- * @param analysis - Analysis string
- * @param fileName - CSV filename (or null)
- * @returns True if sent, false otherwise
- */
-async function sendSlackReport(
-  slackNotifier: SlackNotifier | null,
-  messageTemplate: string,
-  startDate: Date,
-  endDate: Date,
-  csvFilePath: string | null,
-  rowCount: number,
-  analysis: string,
-  fileName: string | null,
-): Promise<boolean> {
-  if (!slackNotifier) {
-    return false;
-  }
-
-  const reportData = {
-    startDate: startDate.toISOString(),
-    endDate: endDate.toISOString(),
-    rowCount,
-    fileName: fileName ?? 'n/a',
-    analysis,
-    timestamp: new Date().toISOString(),
-  };
-
-  await slackNotifier.sendReport(messageTemplate, reportData, csvFilePath);
-  return true;
-}
-
-// ============================================================================
-// Main Function
-// ============================================================================
-
-/**
- * Main script execution function
+ * Main script execution function.
  *
  * Executes Athena queries to monitor TPP messages and generates reports
  * with optional Slack notifications.
@@ -195,7 +36,6 @@ export async function main(script: Core.GOScript): Promise<void> {
   const endDate = config.to ? parseDateTime(config.to) : new Date();
   const startDate = config.from ? parseDateTime(config.from) : hoursAgo(24);
 
-  // Validate date range
   if (startDate >= endDate) {
     throw new Error('Start date must be before end date');
   }
@@ -212,8 +52,6 @@ export async function main(script: Core.GOScript): Promise<void> {
   });
 
   const athenaExecutor = new AthenaQueryExecutor(athenaService, (msg) => script.logger.info(msg));
-
-  // Resolve output path via GOPaths (ensures directory exists, uses /tmp on Lambda)
   const reportsPath = script.paths.resolvePath(config.reportsFolder, Core.GOPathType.OUTPUT) ?? config.reportsFolder;
 
   // Initialize Slack (optional)
@@ -226,15 +64,8 @@ export async function main(script: Core.GOScript): Promise<void> {
   }
 
   try {
-    // Query template from configuration (loaded automatically from config.yaml by GOScript)
-    script.logger.section('Query Template');
-    const queryTemplate = config.athenaQuery;
-    script.logger.info('Query template loaded from configuration');
-
-    // Build query parameters
+    // Build query config and parameters
     const queryParams = buildQueryParams(startDate, endDate);
-
-    // Build Athena config
     const athenaConfig: AthenaQueryConfig = {
       database: config.athenaDatabase,
       catalog: config.athenaCatalog,
@@ -247,14 +78,12 @@ export async function main(script: Core.GOScript): Promise<void> {
     // Execute query
     script.logger.section('Executing Athena Query');
     script.prompt.startSpinner('Running query...');
-    const results = await executeAthenaQuery(athenaExecutor, queryTemplate, athenaConfig, queryParams);
+    const results = await athenaExecutor.executeQuery(config.athenaQuery, athenaConfig, queryParams);
     script.prompt.spinnerStop('Query completed');
-
-    // Convert Athena results to CSV rows
-    const data = convertAthenaResults(results);
 
     // Save and analyze results
     script.logger.section('Processing Results');
+    const data = convertAthenaResults(results);
     const { csvFilePath, fileName, rowCount, analysis } = await saveAndAnalyzeResults(
       data,
       reportsPath,
@@ -287,19 +116,9 @@ export async function main(script: Core.GOScript): Promise<void> {
     script.logger.info(`Total rows: ${rowCount}`);
     script.logger.info(`Slack notification: ${slackSent ? 'SENT' : 'SKIPPED (not configured)'}`);
   } catch (error) {
-    // Attempt to send error to Slack
-    if (slackNotifier) {
-      try {
-        await slackNotifier.sendError('Error during report generation', error instanceof Error ? error : undefined);
-      } catch (slackError) {
-        script.logger.error(
-          `Failed to send error to Slack: ${slackError instanceof Error ? slackError.message : 'Unknown error'}`,
-        );
-      }
-    }
+    await notifySlackError(slackNotifier, error, script);
     throw error;
   } finally {
-    // Cleanup
     athenaService.destroy();
   }
 }
