+++ scripts/send/send-query-dynamodb/src/main.ts
@@ -7,22 +7,20 @@
  * Also prints a clean JSON mapping of results to the console.
  */
 
-import { QueryCommand, DescribeTableCommand } from '@aws-sdk/client-dynamodb';
-import type { QueryCommandInput, AttributeValue, TableDescription } from '@aws-sdk/client-dynamodb';
-import { unmarshall } from '@aws-sdk/util-dynamodb';
+import { DescribeTableCommand } from '@aws-sdk/client-dynamodb';
 import { Core } from '@go-automation/go-common';
 
-import { importPks } from './libs/PkImporter.js';
-import { formatConsoleJson, formatForCsv, formatForText } from './libs/OutputFormatter.js';
-import type { SendQueryDynamodbConfig, OutputFormat } from './types/index.js';
+import { loadPartitionKeys } from './libs/loadPartitionKeys.js';
+import { getSchemaInfo, validateSchemaConfig } from './libs/getSchemaInfo.js';
+import { queryAllKeys } from './libs/queryAllKeys.js';
+import { writeResultsToFile } from './libs/writeResultsToFile.js';
+import { withRetry } from './libs/withRetry.js';
+import { formatConsoleJson } from './libs/OutputFormatter.js';
+import { displayDryRunPreview } from './libs/displayDryRunPreview.js';
+import type { SendQueryDynamodbConfig } from './types/index.js';
 
 /**
- * Result mapping: PK -> array of items
- */
-type ResultMap = Record<string, unknown[]>;
-
-/**
- * Main script execution function
+ * Main script execution function.
  *
  * @param script - The GOScript instance for logging, config, and AWS access
  */
@@ -31,35 +29,14 @@ export async function main(script: Core.GOScript): Promise<void> {
 
   // Step 1: Import PKs
   script.logger.section('Importing Partition Keys');
-  let pks: string[];
-
-  if (config.inputPks) {
-    pks = config.inputPks.map((pk) => pk.trim()).filter((pk) => pk !== '');
-    script.logger.info(`Found ${pks.length} PKs from command line`);
-  } else if (config.inputFile) {
-    const inputPathInfo = script.paths.resolvePathWithInfo(config.inputFile, Core.GOPathType.INPUT);
-
-    script.prompt.startSpinner(`Reading PKs from ${inputPathInfo.path}...`);
-    const imported = await importPks(inputPathInfo.path, {
-      format: config.inputFormat,
-      csvColumn: config.csvColumn,
-      csvDelimiter: config.csvDelimiter,
-    });
-    pks = [...imported];
-    script.prompt.spinnerStop(`Found ${pks.length} unique PKs from file`);
-  } else {
-    throw new Error('Either input.pks or input.file must be provided');
-  }
-
-  // Deduplicate PKs
-  pks = [...new Set(pks)];
+  const pks = await loadPartitionKeys(config, script);
 
   if (pks.length === 0) {
     script.logger.warning('No PKs to process');
     return;
   }
 
-  // Step 2: Preliminary Schema Check
+  // Step 2: Schema check and validation
   script.logger.section('Checking Table Schema');
   script.prompt.startSpinner(`Describing table ${config.tableName}...`);
   const tableDesc = await withRetry(async () => {
@@ -72,121 +49,26 @@ export async function main(script: Core.GOScript): Promise<void> {
     throw new Error(`Could not retrieve description for table ${config.tableName}`);
   }
 
-  const { partitionKey, sortKey } = getSchemaInfo(tableDesc, config.indexName);
+  const schema = getSchemaInfo(tableDesc, config.indexName);
+  validateSchemaConfig(schema, config.tableKey, config.tableSortKey, config.tableSortValue);
 
   script.logger.info(`Target: ${config.indexName ? `Index ${config.indexName}` : 'Base Table'}`);
-  script.logger.info(`Partition Key: ${partitionKey}`);
-  if (sortKey) {
-    script.logger.info(`Sort Key: ${sortKey}`);
+  script.logger.info(`Partition Key: ${schema.partitionKey}`);
+  if (schema.sortKey) {
+    script.logger.info(`Sort Key: ${schema.sortKey}`);
   }
 
-  // Validation
-  if (config.tableKey !== partitionKey) {
-    throw new Error(`Configured table.key (${config.tableKey}) does not match schema partition key (${partitionKey})`);
-  }
-
-  if (sortKey && (!config.tableSortKey || !config.tableSortValue)) {
-    throw new Error(`Table/Index requires a sort key (${sortKey}), but table.sort-key or table.sort-value is missing`);
-  }
-
-  if (config.tableSortKey && config.tableSortKey !== sortKey) {
-    throw new Error(
-      `Configured table.sort-key (${config.tableSortKey}) does not match schema sort key (${sortKey ?? 'none'})`,
-    );
-  }
-
-  // Dry-run preview
+  // Step 3: Dry-run preview
   if (config.dryRun) {
-    script.logger.section('Dry Run Preview');
-    script.logger.info(`Table: ${config.tableName}`);
-    if (config.indexName) script.logger.info(`Index: ${config.indexName}`);
-    script.logger.info(`Partition Key: ${config.tableKey} = [prefix]PK[suffix]`);
-    if (config.tableSortKey) {
-      script.logger.info(`Sort Key: ${config.tableSortKey} = ${config.tableSortValue}`);
-    }
-    const prefix = config.keyPrefix ?? '';
-    const suffix = config.keySuffix ?? '';
-    const previewCount = Math.min(pks.length, 10);
-    pks.slice(0, previewCount).forEach((pk) => script.logger.info(`  Query PK: ${prefix}${pk}${suffix}`));
-    if (pks.length > previewCount) script.logger.info(`  ... and ${pks.length - previewCount} more`);
+    displayDryRunPreview(script, config, pks);
     return;
   }
 
-  // Step 3: Query DynamoDB
+  // Step 4: Query DynamoDB
   script.logger.section('Querying DynamoDB');
-  const resultMap: ResultMap = {};
-  const isRaw = config.outputFormat === 'dynamo-json';
-  const projection = config.outputAttributes
-    ? config.outputAttributes
-        .split(',')
-        .map((a) => a.trim())
-        .filter((a) => a !== '')
-    : undefined;
+  const { resultMap, totalItems } = await queryAllKeys(pks, config, script);
 
-  script.prompt.startSpinner(`Querying ${pks.length} PKs...`);
-
-  let totalItems = 0;
-  const chunkSize = 10;
-
-  for (let i = 0; i < pks.length; i += chunkSize) {
-    const chunk = pks.slice(i, i + chunkSize);
-
-    await Promise.all(
-      chunk.map(async (pk) => {
-        const fullKey = `${config.keyPrefix ?? ''}${pk}${config.keySuffix ?? ''}`;
-        const items: unknown[] = [];
-        let exclusiveStartKey: Record<string, AttributeValue> | undefined;
-
-        do {
-          const keyCondition = config.tableSortKey ? '#pk = :pkVal AND #sk = :skVal' : '#pk = :pkVal';
-
-          const expressionNames: Record<string, string> = { '#pk': config.tableKey };
-          const expressionValues: Record<string, AttributeValue> = { ':pkVal': { S: fullKey } };
-
-          if (config.tableSortKey && config.tableSortValue) {
-            expressionNames['#sk'] = config.tableSortKey;
-            expressionValues[':skVal'] = { S: config.tableSortValue };
-          }
-
-          if (projection) {
-            projection.forEach((attr, idx) => {
-              expressionNames[`#attr${idx}`] = attr;
-            });
-          }
-
-          const input: QueryCommandInput = {
-            TableName: config.tableName,
-            IndexName: config.indexName,
-            KeyConditionExpression: keyCondition,
-            ExpressionAttributeNames: expressionNames,
-            ExpressionAttributeValues: expressionValues,
-            ExclusiveStartKey: exclusiveStartKey,
-            ...(projection && {
-              ProjectionExpression: projection.map((_, idx) => `#attr${idx}`).join(', '),
-            }),
-          };
-
-          const response = await withRetry(async () => script.aws.dynamoDB.send(new QueryCommand(input)));
-
-          if (response.Items) {
-            for (const item of response.Items) {
-              items.push(isRaw ? item : unmarshall(item));
-            }
-          }
-          exclusiveStartKey = response.LastEvaluatedKey;
-        } while (exclusiveStartKey !== undefined);
-
-        resultMap[pk] = items;
-        totalItems += items.length;
-      }),
-    );
-
-    script.prompt.updateSpinner(`Processed ${Math.min(i + chunkSize, pks.length)}/${pks.length} PKs...`);
-  }
-
-  script.prompt.spinnerStop(`Querying completed. Found ${totalItems} items across ${pks.length} PKs.`);
-
-  // Step 4: Save default JSON results to execution output directory
+  // Step 5: Save default JSON results
   script.logger.section('Saving Results');
   const defaultOutputPath = script.paths.resolvePath('results.json', Core.GOPathType.OUTPUT);
   script.prompt.startSpinner('Saving default JSON results mapping...');
@@ -194,137 +76,23 @@ export async function main(script: Core.GOScript): Promise<void> {
   await defaultExporter.export(resultMap);
   script.prompt.spinnerStop(`Default results mapping saved to: ${defaultOutputPath}`);
 
-  // Step 5: Write to custom output file if requested
+  // Step 6: Write to custom output file if requested
   if (config.outputFile) {
     script.logger.section('Writing Custom Output');
     const outputFilePath = script.paths.resolvePath(config.outputFile, Core.GOPathType.OUTPUT);
     script.prompt.startSpinner(`Exporting custom results to ${config.outputFile} (${config.outputFormat})...`);
-
     await writeResultsToFile(outputFilePath, resultMap, config.outputFormat, config.tableKey);
     script.prompt.spinnerStop(`Custom results successfully written to ${config.outputFile}`);
   }
 
-  // Step 6: Summary
+  // Step 7: Summary
   script.logger.section('Summary');
   const withData = Object.values(resultMap).filter((items) => items.length > 0).length;
   script.logger.info(`Total PKs queried: ${pks.length}`);
   script.logger.info(`PKs with results: ${withData}`);
   script.logger.info(`Total items retrieved: ${totalItems}`);
 
-  // Step 6: Console Output (MANDATORY JSON mapping)
+  // Step 8: Console Output (mandatory JSON mapping)
   script.logger.section('Result Mapping');
   console.log(formatConsoleJson(resultMap));
-}
-
-/**
- * Extracts schema info (PK and optionally SK) for the table or specified index
- */
-function getSchemaInfo(
-  table: TableDescription,
-  indexName?: string,
-): { partitionKey: string; sortKey: string | undefined } {
-  let keySchema = table.KeySchema;
-
-  if (indexName) {
-    const gsi = table.GlobalSecondaryIndexes?.find((i) => i.IndexName === indexName);
-    const lsi = table.LocalSecondaryIndexes?.find((i) => i.IndexName === indexName);
-    const index = gsi ?? lsi;
-
-    if (!index) {
-      throw new Error(`Index ${indexName} not found in table description`);
-    }
-    keySchema = index.KeySchema;
-  }
-
-  if (!keySchema) {
-    throw new Error(`No key schema found for ${indexName ? `index ${indexName}` : 'table'}`);
-  }
-
-  const pk = keySchema.find((k) => k.KeyType === 'HASH')?.AttributeName;
-  const sk = keySchema.find((k) => k.KeyType === 'RANGE')?.AttributeName;
-
-  if (!pk) {
-    throw new Error(`Could not find partition key in ${indexName ? `index ${indexName}` : 'table'} schema`);
-  }
-
-  return { partitionKey: pk, sortKey: sk };
-}
-
-/**
- * Retry helper with exponential backoff
- */
-async function withRetry<T>(operation: () => Promise<T>, maxAttempts = 5): Promise<T> {
-  let lastError: unknown;
-  for (let attempt = 1; attempt <= maxAttempts; attempt++) {
-    try {
-      return await operation();
-    } catch (error: unknown) {
-      lastError = error;
-      const retryableNames = [
-        'ProvisionedThroughputExceededException',
-        'LimitExceededException',
-        'InternalServerError',
-        'ServiceUnavailable',
-        'RequestTimeoutException',
-      ];
-
-      const errorObj = error as Record<string, unknown>;
-      const errorName = typeof errorObj['name'] === 'string' ? errorObj['name'] : '';
-      const errorCode = typeof errorObj['code'] === 'string' ? errorObj['code'] : '';
-
-      if (retryableNames.includes(errorName) || errorCode === 'ECONNRESET' || errorCode === 'ETIMEDOUT') {
-        const delay = Math.pow(2, attempt) * 100;
-        await new Promise((resolve) => setTimeout(resolve, delay));
-        continue;
-      }
-      throw error;
-    }
-  }
-  throw lastError;
-}
-
-/**
- * Writes results to file in specified format
- */
-async function writeResultsToFile(
-  path: string,
-  resultMap: ResultMap,
-  format: OutputFormat,
-  pkKey: string,
-): Promise<void> {
-  const items = Object.values(resultMap).flat();
-
-  switch (format) {
-    case 'json':
-    case 'dynamo-json':
-    case 'ndjson': {
-      const isNdjson = format === 'ndjson';
-      const exporter = new Core.GOJSONListExporter<unknown>({
-        outputPath: path,
-        pretty: !isNdjson,
-        indent: 4,
-        jsonl: isNdjson,
-      });
-      await exporter.export(items);
-      break;
-    }
-    case 'csv': {
-      const csvData = formatForCsv(resultMap, pkKey);
-      const exporter = new Core.GOCSVListExporter<Record<string, unknown>>({
-        outputPath: path,
-        includeHeader: true,
-      });
-      await exporter.export(csvData);
-      break;
-    }
-    case 'text': {
-      const text = formatForText(resultMap);
-      const textLines = text.split('\n').filter((line) => line.length > 0);
-      const textExporter = new Core.GOFileListExporter({ outputPath: path });
-      await textExporter.export(textLines);
-      break;
-    }
-    default:
-      throw new Error(`Unsupported output format: ${format as string}`);
-  }
 }
