+++ scripts/send/send-query-dynamodb/src/types/ResultMap.ts
@@ -1,0 +1,4 @@
+/**
+ * Result mapping: partition key to array of DynamoDB items.
+ */
+export type ResultMap = Record<string, unknown[]>;
