+++ scripts/send/send-query-dynamodb/src/types/index.ts
@@ -2,3 +2,4 @@
  * Types barrel file
  */
 export type { SendQueryDynamodbConfig, InputFormat, OutputFormat } from './SendQueryDynamodbConfig.js';
+export type { ResultMap } from './ResultMap.js';
